Cousework/requirements.txt
Package             Version
------------------- -----------
contourpy           1.3.0
cycler              0.12.1
fonttools           4.57.0
importlib_resources 6.5.2
kiwisolver          1.4.7
matplotlib          3.9.4
numpy               2.0.2
packaging           24.2
pandas              2.2.3
pillow              11.1.0
pip                 21.2.4
pyparsing           3.2.3
python-dateutil     2.9.0.post0
pytz                2025.2
PyYAML              6.0.2
scipy               1.13.1
seaborn             0.13.2
setuptools          58.0.4
six                 1.17.0
tzdata              2025.2
zipp                3.21.0
